+++ internal/command/cmd_dbversion.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/korfairo/migratory/internal/gomigrator"
+	"github.com/korfairo/migratory/internal/migrator"
 	"github.com/spf13/cobra"
 )
 
@@ -46,7 +46,7 @@ func getDBVersion(schema, table string) (int64, error) {
 	}()
 
 	ctx := context.Background()
-	migrator, err := gomigrator.New(ctx, db, "postgres", schema, table)
+	migrator, err := migrator.New(ctx, db, "postgres", schema, table)
 	if err != nil {
 		return 0, fmt.Errorf("could not create migrator: %w", err)
 	}
