+++ internal/migrator/migrator_test.go
@@ -1,4 +1,4 @@
-package gomigrator
+package migrator
 
 import (
 	"testing"
